Common up null datasets with minor differences
A minor improvement to the generated code for a few queries.
The type of a dataset may contain an annotation indicating the name of
the original record.  If one null dataset expression contains an
annotation and another doesn't they should be treated as identical.

(Had a minor effect on a couple of queries).

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>