Stop LIMIT SKIP|ONFAIL attributes being ignored
On very rare occasions the SKIP/ONFAIL attribute of a limit could
be ignored.  It occurred when a limit was incorrectly evaluated
using inline code, rather than executed in a child query.  This fix
ensures it is executed in the correct place.

Example was SET(LIMIT(ds, ONFAIL), field).

There were no other examples in any archived queries.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>