FEATURE - hook-based git file reading
Adds code to support reading of files directly from git repositories,
with the ability to specify a revision or tag to use, without needing
to first check out a working copy.

Syntax of a filename is /path/of/repo/.git/{revision}/path/in/repo

The same system can be applied to zip files etc in the future.
eclcc installs the hooks required for this to operate - other components
can do so if appropriate.

Also change how eclcc logging is initialized to avoid git logging from appearing
before the logging level got turned down.

Rather than turning logging off if verbose is not seen, it now turns it off by
default then turns it back on when verbose IS seen.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>