Fix gh-#2065 Replace 'SMC' in ECLWatch/ConfigMgr UI
Existing ECLWatch Permission area shows 'SMC' as feature name.
This fix changes it to better names. Most of the changes are
in environment.xml and other deployment scripts which create
the term into esp.xml. It is decided to rename the 'ESP
Services for SMC' to be 'ESP Services for management console'
and rename the 'Root access to SMC service' to be 'Base
access to ESP services'. This fix also changes the term
'espsmc' to 'eclwatch'.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>