Improved error reporting when LDAP object name exists
If trying to add an LDAP object of type user, group, sudoer (etc) with the
same name as another LDAP object, LDAP reports an 'Already Exists' error
(even if its an object of a different type). But the message we report to
the admin user misleads them to think its an object of the same type.
This change enhances the message to inform the admin that it can be a name
conflict with any LDAP type

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>