Fix problem incorrectly removing COUNTER from PROJECT
There was code in the grammar to check if a COUNTER was used in the
transform of a COUNT PROJECT, and if not report a warning and
convert it to a standard PROJECT.
However the check doesn't work in all cases because it is being
applied to an expression tree that hasn't been normalized.

Move the optimization to hqlopt where the expression tree has
been normalized.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>