Bug #2229 ECLJDBC driver inherits platform's cmake version variables and values.
Although the ECL JDBC driver is a stand-alone JAVA based application, without
any direct dependency on the platform, it will be versioned in parallel with the
platform. In order to achieve this, a simple java class is generated by the CMAKE
configuration process. This "version tracker" class contains final static values
based on the platform's cmake version variables. The entire project is free to use
those fields for version reporting purposes.

Signed-off-by: Rodrigo Pastrana <Rodrigo.Pastrana@lexisnexis.com>