Roxie selftest regression causing core
Recent change to the operator(new) behaviour in roxiemem missed the
case in the selftest code that allocates a heaplet. As a result it
was allocating an invalid healet object and coring in the selftest
code.

Also add some checking to the ccdfile test cases to prevent a core if
current directory is not writable when running the selftest.

Fixes 2307.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>