loading an empty xml stream, created wrong root node type
Noticed in Dali, when loading a 0 length store file.
The IPTreeMaker used to load, created a standard impl. of a IPropertyTree,
when it should have/needed to ask the maker to create the appropriate type.

The result (in Dali) was that it assumed all nodes were of the correct impl.
(CServerRemoteTree) and crashed accessing it via a cast.

NB: only an issue when loading an empty file/stream

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>