Ensure clearAggregate() fields are assigned in order
For a TABLE() that performs an aggregation (no_newaggregate)
the order of the fields in the transform matches
the order of the fields in the target record.  However, for the new
AGGREGATE operator (no_aggregate) they may not match.  If the record
contains variable length fields, and assignments are done out of order
it can lead to invalid memory accesses.

This change ensures assignments are made in field rather than transform
order.

Fixes #2350.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>