ECLPlayground:  Fixed issue when hosted within an iFrame.
ECL Playground fails to load when it is hosted within a frame.  In IDE
this shows a popup window.  In regular IE it was a silent error.

Fixed an issue with result viewer inside a frame (not resizing on
initial load).

Fixed an issue parsing Key=Value params from URL (again related to fames
+ inner).

Fixes gh-2357
Fixes gh-2366

Signed-off-by: Gordon Smith <gordon.smith@lexisnexis.com>