Potential aggregate merge issues if 0 recs on node
A node with no records was being treated as a default aggregate row, i.e.
result of clearAggregate()
This had consequences if the merging function was impacted by default
values.
The example where this was seen (in the ML library), was multiplying by a
field value in the aggregate row, the default value was 0.0, so the end
result was 0.0 if any node had 0 rows to aggregate.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>