Fix gh2460 dropzones not selectable for file upload
On the Upload/Download files page, when multiple dropzones
are set and a user selects the second dropzone, EclWatch
would automatically reselect the default (first) dropzone.
Eclwatch selects the dropzone based on both network address
and dropzone path. When an extra path seperator is at the
end of the dropzone path, the selection fails. The code in
this fix takes care of the extra path seperator.