Fix stack fault if memory callback allocates memory
If a registered IBufferedRowCallback allocated memory while attempting
to free memory in freeBufferedRows() it could recursively call the same
callback resulting in a stackfault.

This change ensure any active callbacks are not called to free memory
allowing another callback to free the necessary memory, or a heap
exception to be generated instead.

Fixes #2446

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>