Fix gh-2480 Can't access archived workunits
In (EclWatch) Search Workunits page, when a user trys to
open an ECL workunit using the workunit ID, WsWorkunits
will try to open the workunit as a non-archived workunit
using WsWuInfo classs. If the ID cannot be found, WsWuInfo
will throw an exception. WsWorkunits should catch the
exception and try to open the workunit as an archived
workunit. The problem here is that WsWorkunits does not
catch the exception correctly.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>