Fix gh-2476 get thor slave log on node getting wrong log
The problem is because the file structure of thor log has
been changed. Under the log folder, there are multiple log
files (instead of one THORMASTER.log). This fix reads the
file name of thor master log from WU XML file and re-creates
the file name for slave log based on the file name of thor
master, as well as the IP and, if any, the port number of
the slave node.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>