Fix gh-2478 wrong order of machines in preflight page
The machines in the preflight page should display in a
correct order. On a default one-node system, looking at
the thor cluster preflight page, the position in the
list of the thor slave and master seems to swap every
time when the page is updated. Two problems have been
found: 1. the machine list inside the preflight request
does not asks for the list to be processed in an order;
2. the preflight code uses a map to store the machine
list and the map re-orders the machines. Those problems
have been fixed now.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>