Optimize (and fix) appending a row to a dataset
Fixes a potential problem when selecting a row from a dataset.  The
row that was selected could be the default row - and if so the selected
row would not be link counted.  So clear the link counted flag if the
row could possibly return a non-link counted default.

Also optimize the code for the case where the row has already been
generated, otherwise the same code could be generated again.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>