Fix gh-2476 get thor slave log on node getting wrong log
The problem is because the file structure of thor log has
been changed. Under the log folder, there are multiple log
files (instead of one THORMASTER.log). This fix reads the
the entry(s) about thor master log from WU XML file. Based
on the thor log entry, EclWatch displays the name of the
thor process(es) and possible slave number for a user to
select. Using the selections, EclWatch retrieves the path
and name of thor slave log, as well as the network address
where the log file can be read. Then, the thor slave log
can be read and displayed.

Since the log entries for eclagent and thormaster are also
changed, this fix modifies ws_workunits code to match with
those changes.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>