Fix gh-2202 Incorrect output from xml output(row)
Added a new workunit method setResultRow() which saves a row and
sets the row count to 1. Call this new method from setResultRaw, which
is called by the workunit shared object to save a row. This fixes the
missing XML output because it now sets a row count, which is needed
by the xml writer

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>