Improve Get Slave Log for Workunit
To get slave log from workunit Details page, a user may
enter a slave number or network address. FireFox disables
Backspace key and Allow keys. When a user types in a wrong
value, the user has to highlight the wrong value and
re-enter. This fix forces those keys enabled so that the
user may use them to modify the value. In addition, this
fix allows a user to get the log file using Enter key as
soon as the value is entered.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>