gh-2614 Regression suite failing on library tests
ecl publish was giving an error if the state of a workunit after the
compile step was completed (rather than compiled).

eclccserver should only set the state to completed on library workunits
if the requested action is 'run'

Related to gh-2614 but not a complete fix for that issue, as non-roxie
cases will still fail until regression suite changed to use publish for
the library queries.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>