Fix gh-2515 Improve behaviour of getWuid() in Roxie
When ECL code calls thorlib.wuid() in a roxie query it gives an internal
error which does not help identify the issue.

Most Roxie queries do not have a unique WUID to return, and an error
seems reasonable (but more specific). A query that was executed via a
dali queue DOES have a workunit, and the wuid should be returned in
such cases.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>