Fix gh-2624 remote.ecl mismatch test in Roxie regression suite
Rework the test so that the output is independent of the number of nodes
in the system, while still testing that the number of results returned
from the ALLNODES() is proportional to the cluster size.

This requires that issue #2656 is fixed if it is to work correctly on
clusters with > 1 channel.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>