Fix gh-2660 CATCH,SKIP in Roxie not aborting upstream activities
CATCH,SKIP in Roxie is not properly aborting the upstream activities when a failure
is caught. This results in some diagnostic tracing about unexpected activity states,
but also means that WHEN() activities above the CATCH are not appropriately triggered.
This causes failures in when6c.ecl regression test.

This patch also corrects the when6c.ecl regression test so that it can be used in the
regression suite without being reported as a failure.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>