Ensure Dali subscription removed on exception
An exception during subscription removal was causing it to be left behind.

Avoid abort() on add/remove subscription exceptions

remove/addd was calling abort() on exception, not sure why, but this was
destroying all objects that other threads may have been about to use.
Also not throwing an exception and simply returning, meant it was likely the
object just being added was now duff.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>