Noreturn in start-stop-daemon
Enabling the HAVE_C_ATTRIBUTE in start-stop-daemon
to correct the behaviour (and suppress warnings)
has the side effect to add another warning, this
time, pointing to a bug in the code.

Since we can't/shouldn't change the source code,
and GCC doesn't recognize -Wno-invalid-noreturn,
the only option is to ignore the first warning.