Add Ubuntu Unity Launcher icon
This adds the unity launcher and a stock LN icon
to Ubuntu desktops. Left-click to go to local
ESP, right-click gives you the options to start,
stop, restart and see the status of the cluster.

You still have to add it to the unity bar yourself,
but searching for "hpcc" on the dash is enough to
find it, drag it and done.