gh-2712, avoid lock on add/del subfiles in DFU
This is a temporary solution until a proper fix can be done.

The short-term proper fix in to allow caching on inactive
transactions via a flag (that should also deprecate the
TempTransaction object in FileServices).

The long-term proper fix is to finally make all access to
the DFS from a single entry point, ie. DFSAccess.