gh-2725 Roxie standalone program delay
If roxie is running without a dali (for example, a standalone program
compiled with -target=roxie), it will pause for 5 seconds at startup
to give the dali connect thread a chance to connect, even though that
thread is never started.

This is a new problem introduced in 3.8.0 rc7.

Fixes gh-2725.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>