Provide a function for estimating allocated row capacity
For hash dedup if the dedup criteria are much smaller than the full row
then it is worth saving just the dedup fields.  However this may cause
extra work, may use more memory if the deduped rows are held in memory
(e.g., by a sort), or if there is little difference in the shrunk size.

To help this last criteria, this function allows thor to determine
whether or not removing some fields is likely to reduce the memory
usage.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>