Remove fixmissing flag
This flag controlling an internal process to the
superfile, arguably on the wrong place, was being
passed around many methods and exposed to the
directory and even FileServices.

The assumption that it should fail on some cases
and only warn on others is wrong. This method
should not attempt to bypass problems in the
DFS, as it'll only silently propagate the original
problem.

The behaviour now is to throw at the slightest
sign of trouble, only warning when the file is
foreign (external Dali) or the number of subfiles
is different. The latter might be an error, too,
but I'll leave it for later consideration.