Allow RIGHT ONLY joins against indexes
Previously JOIN(ds,index,condition,RIGHT ONLY); generated an error
because keyed joins didn't support right only joins.  It also meant
that in some situations (see examples) the error was missed, and an
invalid keyed join was generated.

This change allows that join to occur - but still generates an error
if KEYED() is used in the join condition, or as an attribute of the
join.

Addresses bugzilla #100824

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>