fixes gh-2753 Fixed error generating second pdf file created from single cmake file.
When TARG is being set by the macro RUN_XSTLPROC it is being set on the global scope for the current cmake file instead of in the local scope of the macro. This is causing an issue where the second call to make a pdf which calls RUN_XSLTPROC sees the variable as set if it was set on the first run.

Signed-off-by: Philip Schwartz <philip.schwartz@lexisnexis.com>