Fix gh-2747 empty page from Users link for non-admin
When logged in as a non-administrator user in eclwatch, and
you click on any of the 'Users/Permissions' links (Users,
Groups, Permissions) it returns a blank page. Previously,
the page showed an exception: Access denied, administrators
only.

The exception did not show because XslProcessor cannot be
found for the exception xslt. The XslProcessor of
Cws_accessSoapBinding was replaced by the XslProcessor of
Cws_accessSoapBindingEx, which is incorrect. This fix
removed the the XslProcessor of Cws_accessSoapBindingEx.
This fix also removed the getQualifiedNames() which should
not be there any more.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>