gh-2764 Core dumps possible if mix debug/release .so's
The unicodelib plugin compiled in debug mode can core if used with a jlib
compiled in release mode (and probably the reverse). This is caused by
the different definition of the Atom class in the two modes, which causes
chaos if inline calls to the constructor are made in both modes.

Modern debuggers are capable of displaying the value in an Atom without
the 'DbgReadable' field, so remove it.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>