gh-2675 - Implementation of spilling hash dedup
Allow DEDUP,HASH to deal with datasets larger than memory.

General idea is to spill to buckets when memory full, then process each
bucket in turn, creating more buckets if still too much to deal with.
The impl. streams rows from the input, creating HT entries based on the keys
it's dedupping on. The 1st unmatched row is output, the rest are dedupped.

Each phase creates N in memory buckets. Rows are streamed into the buckets,
the 1st row added is known to be unique and is returned as output from
hashdeup. On a spill request, the bucket is flushed to disk, to be used by
the next phase, rows are still added to the HT, so that the rest of the
stream can have some chance of dedupping, but no rows will be returned as
output. On future spill requests to an already spilled bucket, the HT will
be discarded.

Once all have been filtered/spilt by the existing HT, Any reamining HT is discarded and the buckets are read as input, and the process repeated.

The # of buckets is estimated if possible, based on the # of inputs rows.
The 1st round the # of rows may not be available (unless in meta).
On subsequent rounds (reading back buckets), the total # of rows + # of rows it reached at the spill point are known and are used as the basis for # of buckets in the next round.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>