Re-Enable unittests
Re-enabling the unit test infrastructure, adding new tests
for URI and linking existing tests into it. Some other tests
are not yet linked (aren't cppunit or aren't in a shared lib)
and will be done in the furure.

There is a good documentation on how to extend, and about each
part of the infrastructure. The remote tests are also a good
example on how to create more API/TDD unit tests in the future.