gh-2390 ROLLUP should compare previous input row
ROLLUP was in some cases using the current value of the output row
to compare against the next input row, rather than the previus input
row. This could lead to unexpected behaviour if the fields being
compared were not copied from the inptu to the output.

In Thor, only LOCAL rollup behaves this way, as otherwise it is
transformed to use GROUP.

This commit fixes the Roxie and hthor cases.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>