LDAPConfig to filter out DC entries for DNS
When LDAPConfig queries the domain controller for the DC= baseDC adornments,
if DNS is enabled then those entries are returned in addition to the usual
one (dc=internal,dc=sds). In the past we have gotten lucky and the one
we wanted was returned first, but now that we have raised several controllers
we have found this isn't always the case. This fix filters out the DNS entries
and then selects the first one. It is possible as we go forward that there
may be additional ones returned, in which case we may need to externalize
the baseDC setting to configmanager.
into configmanager

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>