Fix Clang errors
Two errors detected by Clang: returning false where a pointer was
expected, and using modifiers on a list of types.

This last one was probably an error in Clang, but it's a real
issue. We should enhance the dumper to extract the modifiers
before dumping the types and print them.