Customer request to modify SOAPCALL timeout granularity
It was requested by UK LN (insurance) to modify SOAPCALL/HTTPCALL to allow
the timeout to be specified with millisecond granularity (eg TIMEOUT(1.5),
TIMELIMIT(5.5))instead of the current method of only allowing num seconds.

This enhancement modifies the parser to get these values as REAL/DOUBLE,
and 2 new methods getTimeoutMS() and getTimeLimitMS() are added to the
generated helper SO (legacy API left intact). SOAPCALL handler modified to
query the new DOUBLE values if they exist, otherwise gets the legacy integer
ones.
Also fixed a problem in SOAPCALL where it only retried "retries - 1" instead
ir "retries" times, and modified the socket read timeout to the lesser of
the specified timeout and the timeLimit remaining.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>