FIX: gh-2862 Ignore case of web service.method
If the casing of ECL SOAPCALL servicename/method specified in the ECL does
not exactly match the case returned in the response, then no data will be
returned (and its almost impossible for the ECL developer to know why).
Since ECL is not a strongly case sensitive language, it was
agreed to allow mismatched case in that situation.
This fix modifies the XPath parser "match" method to support a "noCase"
flag, also modifies the XMLParser to ignore the case if parsing the
service.method node (controlled by a flag passed in the ctor for the parser.
Also modified the SOAPCALL parser instantiation to pass this flag, which
ensures other users of the XML parser are not affected

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>