Fix gh-2955 Spaces not trimmed from workunit id
Searching for a workunit with an id that ends in a space
leads to an internal error. This fix trims both left spaces
and right spaces of a workunit id input for all of the
methods in ws_workunit. This fix also checks the format
of the id and, if the id contains invalid characters,
displays an error: 'Invalid Workunit ID'.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>