gh-2636 Refactor Preflight code to avoid multiple SSH calls to same node
The exsiting Preflight code has to run two ssh commands for each process, which
is not efficient if multiple processes are running in the same node. This fix
modifies the preflight script and web service to do one ssh command for each
node. 1. analyze preflight request and group processes based on machine node;
2. create a new SSH command for multiple processes; 3. run new preflight script;
4. build preflight response based on the results.

This fix also cleaned the process filters.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>