gh-3009 LOCAL on a keyed join causes Roxie to misbehave
LOCAL on a keyed join is really only meaningful inside an ALLNODES
(typically used with NOROOT indexes). However, as coded, LOCAL would
cause incorrect behaviour if specified in other situations.

This change causes the local to be ignored except in an ALLNODES or
potentially in a child query called from remote code such as an index
filter condition. It's debatable what the correct behaviour in the latter
case is.

It's also debatable whether the compiler should have given a warning or
error when LOCAL is specified in cases that do not make sense.

Fixes gh-3009.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>