Fix gh-2933 - Ensure pid files are removed on reboot
Thor creates some additional pid files in it's instance directory.
If the node is cold booted, these pid files persist and are checked on the next Thor startup, by that time the pids could belond to other processes, in which case Thor can deadlock waiting for an old pid to exit, or in the case of slaves, could kill a random hpcc process believing it was a thorslave.

Ensure pid files are written to /var/run, which is cleared on reboot

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>