Fix gh-3070 superfile deleted when removing its subfiles
In existing EclWatch, removing all subfiles from a superfile
will delete the superfile itself.  This fix adds a checkbox
on Removing Subfiles page before a user deletes subfiles.
The checkbox will be enabled when a user is selected to
delete all of the subfiles. The user may choose to delete
(or not delete) the superfile using the checkbox.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>