FIX gh-3023 Fix error calling MakeAbsoluteSD
MakeAbsoluteSD is a Win32 API, and when it fails, a call to GetLastError
retrieves the error code. However in Linux it is a method call implemented
in permissions.cpp, and in that implemenation it does not "SetLastError"
(a Win32 call).

In this bug, the call to MakeAbsoluteSD failed, and the error message was
built up from a call to GetLastError, which returns 0 since nothing was
set. This fix adds debug logging to the Linux implementation of that call,
and does not call GetLastError unless its a Win32 build

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>