Fix gh-3081 Thor slave not shown when group name not set
The preflight does not see slave node after resizing/renaming
cluster. The problem also happens on Cluster Process page
(before the preflight page). It was due to empty group name
read from getClusterGroupName(). This fix checks the attribute
'nodeGroup', If it is emoty, use the attribute 'name' as group
name.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>