gh-3098 Misleading git tag displayed on internal builds
The git tags on internal developer builds can be misleading, because
of the way the upmerges are being done. When on a branch that is not
a specific closedown branch for a tagged build, the git tag is liable
to find the most recently-tagged release version that has been upmerged
to the current branch, and describe the current revision based on that,
which isn't really very helpful.

Add the expected version number to the filter when looking up git tags,
so that the tag will only be used in the cases where it will be helpful.

Fixes gh-3098

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>