gh-3106 CONTRIBUTORS file updated to reflect requirement for issue ref.
Generating changelogs and generally tracking progress on builds (and working
around githib shortcomings) is going to be a lot easier if we get stricter
about the formatting of commit messages.

Reword the CONTRIBUTORS file to make the use of a github issue reference a
requirement rather than a guideline, and provide git hooks to allow easy
checking.

Fixes gh-3106.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>