Fix gh-3083 - logical rename, not renaming physicals correctly
The rename code was assuming the logical file directory matched the current instance base directory.
Which is not likely to be true, e.g. eclagents vs thor's have different default data directories and the rename context is normally eclagent. So renaming a thor file in eclagent, it will normally mismatch.
If it detected a mismatch, there was some odd legacy code which was effectively taking the 1st two head subdirectories from the existing path and using that as the new head.
The result on a default OSS system, was that renames attempt to rename to /var/lib/<filename>

This commit, deducing the old base directory based on the existing lfn directory.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>