Release Cycle explanation
A brief explanation of the types of releases, with a first attempt to naming
them.

This is the start for a formal definition of our release cycle. We also need
a public document stating the connection between our releases (the new names)
and our versions (to which version they relate) at any moment in time, with
a link to this explanation.

That would give developers and users the ability to how to prioritise bugs and
when to expect fixes, respectively.