gh-3120 Bad error when viewing WU result on WU Details page
When a user views WU result on WU Details page and the user
does not have permission for the file, the existing page shows
that 'This section cannot be dispayed due to an exception.'
and original exception is written into log file. This fix
displays the original exception on the WU Details page.

Fixes gh-3120.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>