gh-2333 Outdated Optional packages still display in EclWatch
Optional resource packages no longer come with hpcc installation
package. This fix adds a cmake option 'USE_RESOURCE' and the
default value is OFF. As the default, ECLWatch will only
display a web link to the hpcc website for resources and will
not show the download link for the optional packages.

Fixes gh-2333.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>